+++ package.json
@@ -16,10 +16,14 @@
     "@types/react-dom": "^16.9.0",
     "@types/react-redux": "^7.1.7",
     "@types/styled-components": "^5.1.14",
+    "framer-motion": "^4.1.17",
+    "lodash": "^4.17.21",
+    "polished": "^4.1.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.0",
     "react-scripts": "4.0.3",
+    "recharts": "^2.1.2",
     "styled-components": "^5.3.1",
     "typescript": "~4.1.5"
   },
